# Avoid some unnecessary re-rendering of CalendarDays