--- a/src/components/CalendarMonth.jsx
+++ b/src/components/CalendarMonth.jsx
@@ -27,7 +27,48 @@ import {
   HORIZONTAL_ORIENTATION,
   VERTICAL_SCROLLABLE,
   DAY_SIZE,
+  BLOCKED_MODIFIER,
 } from '../constants';
+import getPhrase from '../utils/getPhrase';
+import { isSelected } from '../utils/getCalendarDaySettings';
+
+function getAriaLabel(phrases, modifiers, day, ariaLabelFormat) {
+  if (!day) {
+    return '';
+  }
+
+  const {
+    chooseAvailableDate,
+    dateIsUnavailable,
+    dateIsSelected,
+    dateIsSelectedAsStartDate,
+    dateIsSelectedAsEndDate,
+  } = phrases;
+
+  const formattedDate = {
+    date: day.format(ariaLabelFormat),
+  };
+
+  if (modifiers) {
+    if (modifiers.has('selected-start') && dateIsSelectedAsStartDate) {
+      return getPhrase(dateIsSelectedAsStartDate, formattedDate);
+    }
+
+    if (modifiers.has('selected-end') && dateIsSelectedAsEndDate) {
+      return getPhrase(dateIsSelectedAsEndDate, formattedDate);
+    }
+
+    if (isSelected(modifiers) && dateIsSelected) {
+      return getPhrase(dateIsSelected, formattedDate);
+    }
+
+    if (modifiers.has(BLOCKED_MODIFIER)) {
+      return getPhrase(dateIsUnavailable, formattedDate);
+    }
+  }
+
+  return getPhrase(chooseAvailableDate, formattedDate);
+}
 
 const propTypes = forbidExtraProps({
   ...withStylesPropTypes,
@@ -86,7 +127,7 @@ const defaultProps = {
   // i18n
   monthFormat: 'MMMM YYYY', // english locale
   phrases: CalendarDayPhrases,
-  dayAriaLabelFormat: undefined,
+  dayAriaLabelFormat: 'dddd, LL',
   verticalBorderSpacing: undefined,
 };
 
@@ -221,21 +262,23 @@ class CalendarMonth extends React.PureComponent {
           <tbody>
             {weeks.map((week, i) => (
               <CalendarWeek key={i}>
-                {week.map((day, dayOfWeek) => renderCalendarDay({
-                  key: dayOfWeek,
-                  day,
-                  daySize,
-                  isOutsideDay: !day || day.month() !== month.month(),
-                  tabIndex: isVisible && isSameDay(day, focusedDate) ? 0 : -1,
-                  isFocused,
-                  onDayMouseEnter,
-                  onDayMouseLeave,
-                  onDayClick,
-                  renderDayContents,
-                  phrases,
-                  modifiers: modifiers[toISODateString(day)],
-                  ariaLabelFormat: dayAriaLabelFormat,
-                }))}
+                {week.map((day, dayOfWeek) => {
+                  const dayModifiers = modifiers[toISODateString(day)];
+                  return renderCalendarDay({
+                    key: dayOfWeek,
+                    day,
+                    daySize,
+                    isOutsideDay: !day || day.month() !== month.month(),
+                    tabIndex: isVisible && isSameDay(day, focusedDate) ? 0 : -1,
+                    isFocused,
+                    onDayMouseEnter,
+                    onDayMouseLeave,
+                    onDayClick,
+                    renderDayContents,
+                    ariaLabel: getAriaLabel(phrases, dayModifiers, day, dayAriaLabelFormat),
+                    modifiers: dayModifiers,
+                  });
+                })}
               </CalendarWeek>
             ))}
           </tbody>