--- a/src/components/DayPickerSingleDateController.jsx
+++ b/src/components/DayPickerSingleDateController.jsx
@@ -153,6 +153,7 @@ export default class DayPickerSingleDateController extends React.PureComponent {
     this.isTouchDevice = false;
     this.today = moment();
 
+    this.modifiersCache = new Map();
     this.modifiers = {
       today: day => this.isToday(day),
       blocked: day => this.isBlocked(day),
@@ -487,8 +488,19 @@ export default class DayPickerSingleDateController extends React.PureComponent {
     return modifiers;
   }
 
+  getCachedModifiers(modifiersArray) {
+    const cacheKey = modifiersArray.join('!');
+
+    if (!this.modifiersCache.has(cacheKey)) {
+      this.modifiersCache.set(cacheKey, new Set(modifiersArray));
+    }
+
+    return this.modifiersCache.get(cacheKey);
+  }
+
   getModifiersForDay(day) {
-    return new Set(Object.keys(this.modifiers).filter(modifier => this.modifiers[modifier](day)));
+    const modifiers = Object.keys(this.modifiers).filter(modifier => this.modifiers[modifier](day));
+    return this.getCachedModifiers(modifiers);
   }
 
   getStateForNewMonth(nextProps) {
@@ -540,7 +552,7 @@ export default class DayPickerSingleDateController extends React.PureComponent {
           ...days,
           [monthIso]: {
             ...month,
-            [iso]: modifiers,
+            [iso]: this.getCachedModifiers(Array.from(modifiers)),
           },
         };
       }, updatedDaysAfterAddition);
@@ -554,7 +566,7 @@ export default class DayPickerSingleDateController extends React.PureComponent {
         ...updatedDaysAfterAddition,
         [monthIso]: {
           ...month,
-          [iso]: modifiers,
+          [iso]: this.getCachedModifiers(Array.from(modifiers)),
         },
       };
     }
@@ -594,7 +606,7 @@ export default class DayPickerSingleDateController extends React.PureComponent {
           ...days,
           [monthIso]: {
             ...month,
-            [iso]: modifiers,
+            [iso]: this.getCachedModifiers(Array.from(modifiers)),
           },
         };
       }, updatedDaysAfterDeletion);
@@ -608,7 +620,7 @@ export default class DayPickerSingleDateController extends React.PureComponent {
         ...updatedDaysAfterDeletion,
         [monthIso]: {
           ...month,
-          [iso]: modifiers,
+          [iso]: this.getCachedModifiers(Array.from(modifiers)),
         },
       };
     }