--- a/stories/DayPicker.js
+++ b/stories/DayPicker.js
@@ -20,7 +20,7 @@ const TestPrevIcon = () => (
       position: 'absolute',
       width: '40px',
     }}
-    tabindex="0"
+    tabIndex="0"
   >
     Prev
   </div>
@@ -37,7 +37,7 @@ const TestNextIcon = () => (
       right: '24px',
       width: '40px',
     }}
-    tabindex="0"
+    tabIndex="0"
   >
     Next
   </div>
@@ -57,19 +57,20 @@ const TestCustomInfoPanel = () => (
 
 storiesOf('DayPicker', module)
   .add('default', withInfo()(() => (
-    <DayPicker />
+    <DayPicker id="storybook-daypicker" />
   )))
   .add('with custom day size', withInfo()(() => (
-    <DayPicker daySize={50} />
+    <DayPicker id="storybook-daypicker" daySize={50} />
   )))
   .add('single month', withInfo()(() => (
-    <DayPicker numberOfMonths={1} />
+    <DayPicker id="storybook-daypicker" numberOfMonths={1} />
   )))
   .add('3 months', withInfo()(() => (
-    <DayPicker numberOfMonths={3} />
+    <DayPicker id="storybook-daypicker" numberOfMonths={3} />
   )))
   .add('vertical', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       numberOfMonths={2}
       orientation={VERTICAL_ORIENTATION}
     />
@@ -82,6 +83,7 @@ storiesOf('DayPicker', module)
       }}
     >
       <DayPicker
+        id="storybook-daypicker"
         numberOfMonths={12}
         orientation={VERTICAL_SCROLLABLE}
       />
@@ -89,6 +91,7 @@ storiesOf('DayPicker', module)
   )))
   .add('vertical with custom day size', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       numberOfMonths={2}
       orientation={VERTICAL_ORIENTATION}
       daySize={50}
@@ -96,6 +99,7 @@ storiesOf('DayPicker', module)
   )))
   .add('vertical with custom height', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       numberOfMonths={2}
       orientation={VERTICAL_ORIENTATION}
       verticalHeight={568}
@@ -104,6 +108,7 @@ storiesOf('DayPicker', module)
   .add('vertical with DirectionProvider', withInfo()(() => (
     <DirectionProvider direction={DIRECTIONS.RTL}>
       <DayPicker
+        id="storybook-daypicker"
         numberOfMonths={2}
         orientation={VERTICAL_ORIENTATION}
         isRTL
@@ -119,6 +124,7 @@ storiesOf('DayPicker', module)
         }}
       >
         <DayPicker
+          id="storybook-daypicker"
           numberOfMonths={12}
           orientation={VERTICAL_SCROLLABLE}
         />
@@ -127,18 +133,21 @@ storiesOf('DayPicker', module)
   )))
   .add('with custom arrows', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       navPrev={<TestPrevIcon />}
       navNext={<TestNextIcon />}
     />
   )))
   .add('with custom details', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       renderDayContents={day => (day.day() % 6 === 5 ? '😻' : day.format('D'))}
     />
   )))
   .add('vertical with fixed-width container', withInfo()(() => (
     <div style={{ width: '400px' }}>
       <DayPicker
+        id="storybook-daypicker"
         numberOfMonths={2}
         orientation={VERTICAL_ORIENTATION}
       />
@@ -146,6 +155,7 @@ storiesOf('DayPicker', module)
   )))
   .add('with info panel', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       renderCalendarInfo={() => (
         <TestCustomInfoPanel />
       )}
@@ -153,14 +163,16 @@ storiesOf('DayPicker', module)
   )))
   .add('with custom week day format', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       weekDayFormat="ddd"
     />
   )))
   .add('with no animation', withInfo()(() => (
     <DayPicker
+      id="storybook-daypicker"
       transitionDuration={0}
     />
   )))
   .add('noBorder', withInfo()(() => (
-    <DayPicker noBorder />
+    <DayPicker id="storybook-daypicker" noBorder />
   )));