I have been investigating performance and noticed that when selecting a
date, all of the days for all of the months end up re-rendering. I
discovered that there are two causes for this currently:

1. All of the phrases are passed down to the CalendarDay components, and
   the phrases object can be different when selecting dates because of
   code like this, which toggles one translation back and forth:

   https://github.com/airbnb/react-dates/blob/888f6505/src/components/DayPickerRangeController.jsx#L501-L508

2. Similarly, all of the modifiers end up being different by reference
   because we end up creating new Sets for them every time, even if the
   sets are deeply equal.

To address the first issue, I decided to simplify the API of CalendarDay
a little, so that it takes an ariaLabel prop which is a simple string,
instead of a phrases prop. This allows us to prevent the changing of the
phrases object from automatically deoptimizing our leaf components.

And, to address the second issue, I introduced a simple cache for
modifier Sets. This allows the downstream components to perform a simple
referential equality check on them to avoid re-rendering, which we get
for free with PureComponent.

These two optimizations prevent many dates from re-rendering, namely
dates that are blocked and out of range. In storybook on my Macbook pro
(non-production React) this seems to bring down the time it takes to
update DayPickerRangeController when selecting dates from ~150ms to
~110ms. I think this will have a positive effect on some other
operations or other types of datepickers, but I didn't profile them.

It is worth noting that there is still a lot of re-rendering happening,
but that is because the value of ariaLabel ends up changing while you
select dates from:

> Choose Thursday, May 30, 2019 as your check-in date. It’s available.

to:

> Choose Thursday, May 30, 2019 as your check-out date. It’s available.

This seems important for accessibility, and I don't have any good ideas
at this moment of how to avoid this yet.

Before:
![image](https://user-images.githubusercontent.com/195534/58577080-91d56800-81fa-11e9-9408-1978e90a7a4a.png)


After:
![image](https://user-images.githubusercontent.com/195534/58577260-f42e6880-81fa-11e9-85d4-dc59d5b2f0d3.png)


It looks like there might be a good opportunity to optimize DayPickerRangeController#componentWillReceiveProps--this is where the modifiers are all built on every update.

---
Convert CalendarDay's aria-label to aria-labelledby

I've been working to improve performance of react-dates, and I noticed
that on many interactions every single date ends up being re-rendered.
In my previous commit, I partly fixed this by avoiding passing the
always-new phrases object down to CalendarDay, and to memoize modifiers
so they also don't deoptimize the pure components at the edges.

However, this still left one phrase that always changes when
selecting start and end dates which still caused most dates to be
re-rendered.

To avoid this, instead of passing an aria-label to CalendarDay, which
needs to change when switching from selecting a start date to selecting
an end date, we can render a div that contains the phrase for the
currently focused date, and pass a reference to it to be used by
aria-labelledby.

The downside here is that since ids need to be unique per page and there
is no reliable way to auto-generate an id on the server that will be
consistent on the client, I had to add required id props in a number of
places. For folks using components that already require startDateId, we
can generate a new id based on that value, so the friction there should
be minimal.

Testing this using VoiceOver on Safari and the results seem to be just
what we want.

Profiling this in storybook (non-production React) on my Macbook pro
brings the time it takes to update DayPickerRangeController from ~110ms
after my previous commit to ~75ms. (Before both of these commits, we
were looking at ~150ms).

I think there might still be some opportunity in preventing unnecessary
re-renders of CalendarMonth, but most of the remaining opportunity looks
like it might be in optimizing how modifiers are built in places like
DayPickerRangeController#componentWillReceiveProps.


After:
![image](https://user-images.githubusercontent.com/195534/58595065-518adf80-8224-11e9-93e2-8b3a2b539431.png)