@lencioni 2019-05-29
If this approach looks good, I'll update the tests.