@erin-doyle 2019-05-29
I just want to put a plug in for also making the relevant changes to `DayPickerSingleDateController` that you've made to `DayPickerRangeController`.